fix: update print_integer to handle call _printf("%d", INT_MIN);

+++ print_d.c
@@ -8,7 +8,7 @@
  */
 int printf_integer(va_list val)
 {
-	int num = va_arg(val, int);
+	long num = va_arg(val, int);
 	char buffer[20];
 	int i, len = 0, digitCount = 0;
